+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+- Searching follows looks for substring matches [#670](https://github.com/rokwire/core-building-block/issues/670)
+
 ### Added
 - Support following accounts [#667](https://github.com/rokwire/core-building-block/issues/667)